cerberushost/requirements.txt
Django==1.4.3
django-allauth==0.9.0
django-openid-auth==0.4
httplib2==0.7.7
oauth2==1.5.211
python-openid==2.2.5
requests==1.1.0
djangorestframework==2.1.17




